backend/requirements.txt
aiofiles==25.1.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
bcrypt==3.2.2
cffi==2.0.0
click==8.3.1
ecdsa==0.19.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
pip==26.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
redis==7.3.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzlocal==5.3.1
uvicorn==0.41.0
pytest
httpx
